requirements.txt
Flask==3.0.0
flask-cors==4.0.0
rembg==2.0.57
Pillow==10.1.0
